Pull request 6188 android

+++ libraries/model/src/model/Material.h
@@ -315,7 +315,7 @@ public:
 
         glm::vec2 _spare{ 0.0f };
 
-        uint32_t _key{ 0 }; // a copy of the materialKey
+        float _key{ 0 }; // a copy of the materialKey
 
         // for alignment beauty, Material size == Mat4x4
